moved proto files and add routing adapter
