+++ build.rs
@@ -13,6 +13,7 @@ fn main() {
         "src/simulation/io/proto/types/network.proto",
         "src/simulation/io/proto/types/population.proto",
         "src/simulation/io/proto/types/vehicles.proto",
+        "src/external_services/routing/routing.proto",
     ];
 
     // tell cargo to rerun this build script if any of the proto files change
