+++ src/lib.rs
@@ -6,5 +6,7 @@ pub mod experiments;
 pub mod simulation;
 
 mod dvrp;
+pub mod external_services;
+pub mod generated;
 #[cfg(test)]
 mod test_utils;
