+++ src/simulation/engines/activity_engine.rs
@@ -1,5 +1,5 @@
+use crate::generated::events::Event;
 use crate::simulation::config::Config;
-use crate::simulation::io::proto::events::Event;
 use crate::simulation::messaging::events::EventsPublisher;
 use crate::simulation::population::InternalPerson;
 use crate::simulation::time_queue::{EndTime, TimeQueue};
