+++ src/simulation/io/proto/mod.rs
@@ -2,245 +2,3 @@ pub mod proto_network;
 pub mod proto_population;
 pub mod proto_vehicles;
 pub mod xml_events;
-
-use crate::simulation::io::proto::general::attribute_value::Type;
-use crate::simulation::io::proto::general::AttributeValue;
-use crate::simulation::io::proto::network::{Link, Network, Node};
-use crate::simulation::io::xml::attributes::IOAttribute;
-use prost::Message;
-use std::fs;
-use std::fs::File;
-use std::io::{BufReader, ErrorKind, Read, Seek, Write};
-use std::marker::PhantomData;
-use std::path::Path;
-use tracing::info;
-
-// Include the `messages` module, which is generated from messages.proto
-pub mod events {
-    include!(concat!(env!("OUT_DIR"), "/events.rs"));
-}
-pub mod ids {
-    include!(concat!(env!("OUT_DIR"), "/ids.rs"));
-}
-
-pub mod network {
-    include!(concat!(env!("OUT_DIR"), "/network.rs"));
-}
-pub mod population {
-    include!(concat!(env!("OUT_DIR"), "/population.rs"));
-}
-pub mod vehicles {
-    include!(concat!(env!("OUT_DIR"), "/vehicles.rs"));
-}
-
-pub mod general {
-    include!(concat!(env!("OUT_DIR"), "/general.rs"));
-}
-
-pub fn read_from_file<T: Message + Default>(path: &Path) -> T {
-    info!("Loading message from file at: {path:?}");
-    let mut reader = File::open(path).unwrap_or_else(|_| panic!("Could not open File at {path:?}"));
-
-    let mut bytes = Vec::new();
-    reader
-        .read_to_end(&mut bytes)
-        .unwrap_or_else(|_| panic!("Could not read File at {path:?}"));
-    let wire_type = T::decode(bytes.as_slice()).expect("Failed to decode file contents");
-
-    info!("Finished loading message from file at: {path:?}");
-    wire_type
-}
-
-pub fn write_to_file<T: Message>(message: T, path: &Path) {
-    info!("Starting to write message to file: {path:?}");
-    let bytes = message.encode_to_vec();
-
-    // Create the file and all necessary directories
-    // this doesn't cover some edge cases, but this will do for now
-    //let path = Path::new(file_path);
-    let prefix = path.parent().unwrap();
-    fs::create_dir_all(prefix).unwrap();
-    let mut file =
-        File::create(path).unwrap_or_else(|_| panic!("Failed to create file at: {path:?}"));
-    file.write_all(&bytes)
-        .unwrap_or_else(|_| panic!("Failed to write bytes to file at: {path:?}"));
-    info!("Finished writing message to file: {path:?}");
-}
-
-pub fn read_delimiter<R>(reader: &mut BufReader<R>) -> Option<usize>
-where
-    R: Read + Seek,
-{
-    // read the delimiter of the message. Prost says delimiter is between 1 and 10 bytes
-    // so, read the first 10 bytes of the buffer
-    let mut delim_buffer: [u8; 10] = [0; 10];
-    // this could crash
-    match reader.read_exact(&mut delim_buffer) {
-        Ok(_) => {} // go on.
-        Err(e) => match e.kind() {
-            ErrorKind::UnexpectedEof => return None,
-            _ => {
-                panic!("Error while reading file: {}", e);
-            }
-        },
-    };
-
-    let delimiter =
-        prost::decode_length_delimiter(delim_buffer.as_slice()).expect("error reading delimiter");
-
-    // since the delimiter is a varint figure out how many bytes the delimiter was actually taking
-    // up in the buffer. Set the buffers position to the first byte after the delimiter, which
-    // should be the start of the TimeStep message
-    let delim_encoded_len = prost::encoding::encoded_len_varint(delimiter as u64) as i64;
-    let offset = delim_encoded_len - (delim_buffer.len() as i64);
-    reader
-        .seek_relative(offset)
-        .expect("Seeking relative failed");
-
-    Some(delimiter)
-}
-
-pub struct MessageIter<T, R>
-where
-    T: Message + Default,
-    R: Read + Seek,
-{
-    type_marker: PhantomData<T>,
-    internal_reader: BufReader<R>,
-}
-
-impl<T, R> Iterator for MessageIter<T, R>
-where
-    T: Message + Default,
-    R: Read + Seek,
-{
-    type Item = T;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if let Some(delimiter) = read_delimiter(&mut self.internal_reader) {
-            let mut bytes: Vec<u8> = vec![0; delimiter];
-            self.internal_reader
-                .read_exact(&mut bytes)
-                .expect("Failed to read exact from buffer");
-            let message = T::decode(bytes.as_slice()).expect("Failed to decode message");
-            Some(message)
-        } else {
-            None
-        }
-    }
-}
-
-impl<T, R> MessageIter<T, R>
-where
-    T: Message + Default,
-    R: Read + Seek,
-{
-    pub fn new(reader: R) -> Self {
-        Self {
-            type_marker: Default::default(),
-            internal_reader: BufReader::new(reader),
-        }
-    }
-}
-
-impl AttributeValue {
-    pub(crate) fn new_int(value: i64) -> Self {
-        AttributeValue {
-            r#type: Some(Type::IntValue(value)),
-        }
-    }
-
-    pub(crate) fn new_string(value: String) -> Self {
-        AttributeValue {
-            r#type: Some(Type::StringValue(value)),
-        }
-    }
-
-    pub(crate) fn new_double(value: f64) -> Self {
-        AttributeValue {
-            r#type: Some(Type::DoubleValue(value)),
-        }
-    }
-
-    pub(crate) fn new_bool(value: bool) -> Self {
-        AttributeValue {
-            r#type: Some(Type::BoolValue(value)),
-        }
-    }
-
-    pub fn as_int(&self) -> i64 {
-        match self.r#type.as_ref().unwrap() {
-            Type::IntValue(value) => *value,
-            _ => panic!("Expected int, got {:?}", self),
-        }
-    }
-
-    pub fn as_string(&self) -> String {
-        match self.r#type.as_ref().unwrap() {
-            Type::StringValue(value) => value.clone(),
-            _ => panic!("Expected string, got {:?}", self),
-        }
-    }
-
-    pub fn as_double(&self) -> f64 {
-        match self.r#type.as_ref().unwrap() {
-            Type::DoubleValue(value) => *value,
-            _ => panic!("Expected double, got {:?}", self),
-        }
-    }
-
-    pub fn as_bool(&self) -> bool {
-        match self.r#type.as_ref().unwrap() {
-            Type::BoolValue(value) => *value,
-            _ => panic!("Expected bool, got {:?}", self),
-        }
-    }
-
-    pub fn from_io_attr(attr: IOAttribute) -> AttributeValue {
-        match attr.class.as_str() {
-            "java.lang.String" => AttributeValue::new_string(attr.value),
-            "java.lang.Double" => AttributeValue::new_double(attr.value.parse().unwrap()),
-            "java.lang.Integer" => AttributeValue::new_int(attr.value.parse().unwrap()),
-            "java.lang.Boolean" => AttributeValue::new_bool(attr.value.parse().unwrap()),
-            _ => panic!("Unsupported attribute class: {}", attr.class),
-        }
-    }
-}
-
-impl Network {
-    pub fn from(network: &crate::simulation::network::Network) -> Self {
-        info!("Converting Network into wire format");
-        let nodes: Vec<_> = network
-            .nodes()
-            .iter()
-            .map(|n| Node {
-                id: n.id.internal(),
-                x: n.x,
-                y: n.y,
-                partition: n.partition,
-                cmp_weight: n.cmp_weight,
-            })
-            .collect();
-        let links: Vec<_> = network
-            .links()
-            .iter()
-            .map(|l| Link {
-                id: l.id.internal(),
-                from: l.from.internal(),
-                to: l.to.internal(),
-                length: l.length,
-                capacity: l.capacity,
-                freespeed: l.freespeed,
-                permlanes: l.permlanes,
-                modes: l.modes.iter().map(|id| id.internal()).collect(),
-                partition: l.partition,
-            })
-            .collect();
-
-        Network {
-            nodes,
-            links,
-            effective_cell_size: network.effective_cell_size(),
-        }
-    }
-}
