+++ src/simulation/io/proto/proto_network.rs
@@ -1,13 +1,52 @@
+use crate::generated::network::{Link, Node};
 use crate::simulation::network::Network;
 use std::path::Path;
+use tracing::info;
 
 pub fn load_from_proto(path: &Path) -> Network {
-    let wire_net: crate::simulation::io::proto::network::Network =
-        crate::simulation::io::proto::read_from_file(path);
+    let wire_net: crate::generated::network::Network = crate::generated::read_from_file(path);
     Network::from(wire_net)
 }
 
 pub fn write_to_proto(network: &Network, path: &Path) {
-    let wire_network = crate::simulation::io::proto::network::Network::from(network);
-    crate::simulation::io::proto::write_to_file(wire_network, path);
+    let wire_network = crate::generated::network::Network::from(network);
+    crate::generated::write_to_file(wire_network, path);
+}
+
+impl crate::generated::network::Network {
+    pub fn from(network: &crate::simulation::network::Network) -> Self {
+        info!("Converting Network into wire format");
+        let nodes: Vec<_> = network
+            .nodes()
+            .iter()
+            .map(|n| Node {
+                id: n.id.internal(),
+                x: n.x,
+                y: n.y,
+                partition: n.partition,
+                cmp_weight: n.cmp_weight,
+            })
+            .collect();
+        let links: Vec<_> = network
+            .links()
+            .iter()
+            .map(|l| Link {
+                id: l.id.internal(),
+                from: l.from.internal(),
+                to: l.to.internal(),
+                length: l.length,
+                capacity: l.capacity,
+                freespeed: l.freespeed,
+                permlanes: l.permlanes,
+                modes: l.modes.iter().map(|id| id.internal()).collect(),
+                partition: l.partition,
+            })
+            .collect();
+
+        crate::generated::network::Network {
+            nodes,
+            links,
+            effective_cell_size: network.effective_cell_size(),
+        }
+    }
 }
