+++ src/simulation/io/proto/xml_events.rs
@@ -10,9 +10,9 @@ use xml::attribute::OwnedAttribute;
 use xml::reader::XmlEvent;
 use xml::EventReader;
 
+use crate::generated::events::event::Type;
+use crate::generated::events::Event;
 use crate::simulation::id::Id;
-use crate::simulation::io::proto::events::event::Type;
-use crate::simulation::io::proto::events::Event;
 use crate::simulation::messaging::events::EventsSubscriber;
 use crate::simulation::network::Link;
 use crate::simulation::population::InternalPerson;
