+++ src/simulation/network/mod.rs
@@ -196,8 +196,8 @@ impl From<IONetwork> for Network {
     }
 }
 
-impl From<crate::simulation::io::proto::network::Network> for Network {
-    fn from(value: crate::simulation::io::proto::network::Network) -> Self {
+impl From<crate::generated::network::Network> for Network {
+    fn from(value: crate::generated::network::Network) -> Self {
         let mut result = Network::new();
         result.set_effective_cell_size(value.effective_cell_size);
         for wn in &value.nodes {
